move filters inside superfeed module
